+++ libafl/src/executors/mod.rs
@@ -15,6 +15,13 @@ pub mod timeout;
 #[cfg(any(unix, feature = "std"))]
 pub use timeout::TimeoutExecutor;
 
+/// The essential hooks for inproc executors
+/// hooks are the code that are run before and after each executor run
+#[cfg(unix)]
+pub mod inprocess_hooks_unix;
+#[cfg(windows)]
+pub mod inprocess_hooks_win;
+
 #[cfg(all(feature = "std", feature = "fork", unix))]
 pub mod forkserver;
 #[cfg(all(feature = "std", feature = "fork", unix))]
@@ -115,6 +122,23 @@ pub trait HasObservers: UsesObservers {
 
     /// Get the linked observers (mutable)
     fn observers_mut(&mut self) -> &mut Self::Observers;
+}
+
+/// The hooks that runs before and after running the target
+pub trait ExecutorHooks {
+    /// The pre_exec hook that is called before running a target.
+    #[allow(clippy::unused_self)]
+    fn pre_run_target<E, EM, I, S, Z>(
+        &self,
+        _executor: &E,
+        _fuzzer: &mut Z,
+        _state: &mut S,
+        _mgr: &mut EM,
+        _input: &I,
+    );
+
+    /// The post_exec hooks that runs before and after running the target
+    fn post_run_target(&self);
 }
 
 /// An executor takes the given inputs, and runs the harness/target.
